// src/app/model/profile.mode.ts
export interface Profile {
    user:
    {
    country: string;
    email: string;
    gender: string;
    name: string;
    phone: string;
    surname: string;
    }
    wedding: 
    {
      bride: string;
      c_gender: string;
      couple: string;
      days_Till_wed: number;
      groom: string;
      location: string;
      maxBudget: number;
      minBudget: number;
      wed_date: string;
      wed_type: string;
    };
  }
  